web/requirements.txt
Flask==2.3.3
Flask-Babel==4.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Babel==2.12.1
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
boto3==1.34.0 
Flask-Caching==2.0.2
Pillow==11.3.0
psycopg2-binary==2.9.9
Flask-Session==0.5.0